make: Add a wrapper script to find & run the tt-smi python
Add a wrapper script that finds the tt-smi python interpreter using
pipx.

This is not fundamentally different to the current logic, but it has
more/better error checking, so the user shouldn't be left with confusing
errors.

We could not use a wrapper script, but do all the same logic inside
make. But that has the downside that it can only be used via make.

Using a wrapper script means a user can run it manually if they want to
debug/inspect something about the python environment.