Add a script to build an image
This uses debootstrap to build an image, and perform customisations such
as adding a passwordless user for easy login (username: debian), setting
up dhcp networking, and disabling services that don't make sense.

make, gcc, libc, python, openssh server and vim are installed.

The size comes out to be ~900MB. No effort to minimise it has been made
at this stage. The use of a qcow2 container could be revisited, there's
not much to be gained there.