WIP: Semi working Rust translation of boot.py
This is a translation of boot.py into Rust, using luwen directly rather
than going via pyluwen.

I've been able to get OpenSBI booted using this, but for some reason
Linux doesn't boot. I think the next step is to trace the writes that
the python and Rust versions are doing to confirm they are the same.

It also doesn't reset the card, that has to be done manually first. That
will need to be fixed.