Rust translation of boot.py
This is a translation of boot.py into Rust, using luwen directly rather
than going via pyluwen.

It doesn't reset the card, that has to be done manually first. That will
need to be fixed.

Use git tag for luwen version

I missed that there is a git tag pointing at the version of luwen we
need. Use it.