tt-bh-linux: interrupt x280 after every processed descriptor in virtqueue
Previously we were interrupting the x280 after processing every descriptor in
the virtqueue. While this results in less cpu load on the x280 and faster speeds
(in tests like iperf for eg, we see transfer speeds jump from 200->900 Mb/s), this
sometimes results in the block device waiting too long for responses for it's
requests. The block device makes a ton of small (4096 byte) requests. This is visible
to the user as though the process doing I/O has completely hung (it actually hasn't)
and SIGINTs sent to the process are ineffective too.

To make the user's interactive experience better, we don't wait for every descriptor
to be processed, instead we interrupt the x280 after every request has been processed.

Maybe as a middle ground, we could interrupt after n requests have been processed.