misc: remove support for booting Linux all L2CPU tiles
In order to align with the upstream device tree series, remove the misc
directory which includes dts files files needed for booting Linux on all
four L2CPU tiles at the same time. This is outside the scope for the
KernelCI branch of tt-bh-linux. It can be added back later if necessary.