Move blackhole_defconfig from the linux repo to tt-bh-linux
For this KernelCI branch of tt-bh-linux, the linux kernel branch with
the upstream dts series is used. This does not contain the patch that
adds the blackhole defconfig. Move it to the tt-bh-linux repo for now so
that make build_linux will continue to work.

There are longer term plans to eliminate the need for NONPORTABLE for
the HVC SBI console. This would mean that eventually the upstream
defconfig would work for Blackhole.