console: don't do anything in ack_interrupt method
Originally there was some functionality in this method to
unset the interrupt register bit. That was removed cause it was
happening way too late and causing too many spurious interrupts

Now what's happening is that the interrupt_status register's value
is being set to 0 before the virtio_mmio device's interupt handler
is able to complete. So the interrupt handler thinks the interrupt
is not of type VIRTIO_MMIO_INT_VRING and doesn't process any
virtqueue descriptors.

So just completely remove all the functionality from this function.